Updated Editor UI with groundwork for adding animation points

+++ engine/Common/Common.hpp
@@ -4,7 +4,7 @@
 
 using byte_vec4 = glm::tvec4< uint8_t >;
 
-constexpr float TARGET_FPS = 144.0f;
+constexpr int32_t TARGET_FPS = 144;
 constexpr float TARGET_TIME = 1.0f / TARGET_FPS; // seconds
 constexpr int32_t WIDTH = 1920;
 constexpr int32_t HEIGHT = 1080;
